requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
boto==2.49.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
eval_type_backport==0.2.0
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hume==0.6.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==2.0.0
opencv-python==4.10.0.84
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
typing_extensions==4.12.2
urllib3==1.26.19
websockets==12.0
Werkzeug==3.0.3
zipp==3.19.2